# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SP Manager API",
    "version": "1.0.0",
    "description": "API completa para gestionar y documentar Stored Procedures de SQL Server."
  },
  "servers": [
    {
      "url": "https://sp-manager-demo.vercel.app",
      "description": "Servidor de Producción / Demo"
    }
  ],
  "paths": {
...
    "/api/projects/{projectName}/sps": {
      "get": {
        "tags": ["Projects"],
        "summary": "Obtener SPs de un proyecto",
        "parameters": [
          {
            "name": "projectName",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Nombre del proyecto"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de SPs del proyecto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SPMetadata" }
                }
              }
            }
          }
        }
      }
    },
    "/api/sps": {
      "get": {
        "tags": ["Stored Procedures"],
        "summary": "Buscar stored procedures",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Término de búsqueda"
          },
          {
            "name": "db",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Nombre de la base de datos"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de SPs encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/StoredProcedureCore" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "StoredProcedureCore": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "schema": { "type": "string" },
          "database": { "type": "string" },
          "objectId": { "type": "number" }
        }
      },
      "SPMetadata": {
        "type": "object",
        "properties": {
           "spName": { "type": "string" },
           "description": { "type": "string" },
           "projectReferences": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}